+++ src/components/ui/Select/Birthday/Month/index.tsx
@@ -1,5 +1,5 @@
 import { useMemo, FC } from 'react';
-import { SelectContent, SelectItem } from '../select';
+import { SelectContent, SelectItem } from '../../select';
 
 type MonthProps = {
   startMonth?: number;
